Time: 0 ms (100%), Space: 89.4 MB (62.32%) - LeetHub

+++ 2270-number-of-ways-to-split-array/2270-number-of-ways-to-split-array.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int waysToSplitArray(vector<int>& nums) {
         int res=0;
-        long total=0,left=0;
+        long long total=0,left=0;
         for(int i=0;i<nums.size();i++)
             total+=nums[i];
         for(int i=0;i<nums.size()-1;i++){
